A first step towards decoupling the header files. There is still a significant amount of work left to do.
--HG--
rename : doc/source/special_matrices.rst => doc/source/matrices.rst
rename : doc/source/special_matrices/deterministic.rst => doc/source/matrices/deterministic.rst
rename : doc/source/special_matrices/random.rst => doc/source/matrices/random.rst
rename : examples/special_matrices/Cauchy.cpp => examples/matrices/Cauchy.cpp
rename : examples/special_matrices/CauchyLike.cpp => examples/matrices/CauchyLike.cpp
rename : examples/special_matrices/Circulant.cpp => examples/matrices/Circulant.cpp
rename : examples/special_matrices/Diagonal.cpp => examples/matrices/Diagonal.cpp
rename : examples/special_matrices/DiscreteFourier.cpp => examples/matrices/DiscreteFourier.cpp
rename : examples/special_matrices/Hankel.cpp => examples/matrices/Hankel.cpp
rename : examples/special_matrices/HermitianUniformSpectrum.cpp => examples/matrices/HermitianUniformSpectrum.cpp
rename : examples/special_matrices/Hilbert.cpp => examples/matrices/Hilbert.cpp
rename : examples/special_matrices/Identity.cpp => examples/matrices/Identity.cpp
rename : examples/special_matrices/Legendre.cpp => examples/matrices/Legendre.cpp
rename : examples/special_matrices/NormalUniformSpectrum.cpp => examples/matrices/NormalUniformSpectrum.cpp
rename : examples/special_matrices/OneTwoOne.cpp => examples/matrices/OneTwoOne.cpp
rename : examples/special_matrices/Ones.cpp => examples/matrices/Ones.cpp
rename : examples/special_matrices/PSFW.cpp => examples/matrices/PSFW.cpp
rename : examples/special_matrices/Toeplitz.cpp => examples/matrices/Toeplitz.cpp
rename : examples/special_matrices/Uniform.cpp => examples/matrices/Uniform.cpp
rename : examples/special_matrices/Walsh.cpp => examples/matrices/Walsh.cpp
rename : examples/special_matrices/Wilkinson.cpp => examples/matrices/Wilkinson.cpp
rename : examples/special_matrices/Zeros.cpp => examples/matrices/Zeros.cpp
rename : include/elemental/special_matrices.hpp => include/elemental/matrices.hpp
rename : include/elemental/special_matrices/Cauchy.hpp => include/elemental/matrices/Cauchy.hpp
rename : include/elemental/special_matrices/CauchyLike.hpp => include/elemental/matrices/CauchyLike.hpp
rename : include/elemental/special_matrices/Circulant.hpp => include/elemental/matrices/Circulant.hpp
rename : include/elemental/special_matrices/Diagonal.hpp => include/elemental/matrices/Diagonal.hpp
rename : include/elemental/special_matrices/DiscreteFourier.hpp => include/elemental/matrices/DiscreteFourier.hpp
rename : include/elemental/special_matrices/Hankel.hpp => include/elemental/matrices/Hankel.hpp
rename : include/elemental/special_matrices/HermitianUniformSpectrum.hpp => include/elemental/matrices/HermitianUniformSpectrum.hpp
rename : include/elemental/special_matrices/Hilbert.hpp => include/elemental/matrices/Hilbert.hpp
rename : include/elemental/special_matrices/Identity.hpp => include/elemental/matrices/Identity.hpp
rename : include/elemental/special_matrices/Legendre.hpp => include/elemental/matrices/Legendre.hpp
rename : include/elemental/special_matrices/NormalUniformSpectrum.hpp => include/elemental/matrices/NormalUniformSpectrum.hpp
rename : include/elemental/special_matrices/OneTwoOne.hpp => include/elemental/matrices/OneTwoOne.hpp
rename : include/elemental/special_matrices/Ones.hpp => include/elemental/matrices/Ones.hpp
rename : include/elemental/special_matrices/Toeplitz.hpp => include/elemental/matrices/Toeplitz.hpp
rename : include/elemental/special_matrices/Uniform.hpp => include/elemental/matrices/Uniform.hpp
rename : include/elemental/special_matrices/Walsh.hpp => include/elemental/matrices/Walsh.hpp
rename : include/elemental/special_matrices/Wilkinson.hpp => include/elemental/matrices/Wilkinson.hpp
rename : include/elemental/special_matrices/Zeros.hpp => include/elemental/matrices/Zeros.hpp
rename : include/elemental/special_matrices_decl.hpp => include/elemental/matrices_decl.hpp
rename : include/elemental/special_matrices_impl.hpp => include/elemental/matrices_impl.hpp

+++ tests/blas-like/Gemm.cpp
@@ -7,6 +7,7 @@
    http://opensource.org/licenses/BSD-2-Clause
 */
 #include "elemental.hpp"
+#include "elemental/matrices/Uniform.hpp"
 using namespace std;
 using namespace elem;
 
@@ -226,8 +227,7 @@ main( int argc, char* argv[] )
                  << "Testing with doubles:\n"
                  << "---------------------" << endl;
         }
-        TestGemm<double>
-        ( print, orientA, orientB, m, n, k, (double)3, (double)4, g );
+        TestGemm<double>( print, orientA, orientB, m, n, k, 3., 4., g );
 
         if( commRank == 0 )
         {
