+++ tests/blas-like/Hemm.cpp
@@ -7,6 +7,7 @@
    http://opensource.org/licenses/BSD-2-Clause
 */
 #include "elemental.hpp"
+#include "elemental/matrices/HermitianUniformSpectrum.hpp"
 using namespace std;
 using namespace elem;
 
@@ -107,8 +108,7 @@ main( int argc, char* argv[] )
                  << "Testing with doubles:                 \n"
                  << "--------------------------------------" << endl;
         }
-        TestHemm<double>
-        ( print, side, uplo, m, n, (double)3, (double)4, g );
+        TestHemm<double>( print, side, uplo, m, n, 3., 4., g );
 
         if( commRank == 0 )
         {
