+++ tests/blas-like/Trmm.cpp
@@ -7,6 +7,7 @@
    http://opensource.org/licenses/BSD-2-Clause
 */
 #include "elemental.hpp"
+#include "elemental/matrices/Uniform.hpp"
 using namespace std;
 using namespace elem;
 
@@ -104,8 +105,7 @@ main( int argc, char* argv[] )
                  << "Testing with doubles:\n"
                  << "---------------------" << endl;
         }
-        TestTrmm<double>
-        ( print, side, uplo, orientation, diag, m, n, (double)3, g );
+        TestTrmm<double>( print, side, uplo, orientation, diag, m, n, 3., g );
 
         if( commRank == 0 )
         {
